Rename remote response files
