--- app/src/main/java/br/edu/jonathangs/pokmontcgdeveloper/network/data/Error.kt
@@ -1,4 +1,0 @@
-package br.edu.jonathangs.pokmontcgdeveloper.network.data
-
-
-data class Error(val status: Int, val error: String)
