+++ app/src/main/java/br/edu/jonathangs/pokmontcgdeveloper/network/data/ErrorResponse.kt
@@ -1,0 +1,4 @@
+package br.edu.jonathangs.pokmontcgdeveloper.network.data
+
+
+data class ErrorResponse(val status: Int, val error: String)
